BetaModels.jl Documentation

API

Distributions.Normal — Method

Normal(B::Beta{Float64})

Return Normal approximation to a Beta.

If $B\sim\text{Beta}(a,b)$ then

\[B\approx\text{Normal}\left(\frac{a}{a+b}, \frac{ab}{(a+b)^2(a+b+1)}\right)\]

The approximation improves as $a\approx b$ and for large values of $a$ and $b$.

An improvement can be obtained by transforming the parameter (not implemented), Wise's transformation.

Base.:- — Method

Base.:-(X::Normal{<:Real}, Y::Normal{<:Real})

Subtract a normal random variable Y from another normal random variable X according to

\[\text{Normal}(\mu, \sigma^2) - \text{Normal}(\nu, \omega^2) \sim \text{Normal}(\mu-\nu, \sigma^2+\omega^2)\]

Base.:- — Method

Base.:-(X::Normal{<:Real})

Return the negative of a Normal random variable, i.e. $X \sim \text{Normal}(\mu,\sigma^2)$ then return $-X \sim \text{Normal}(-\mu,\sigma^2)$.

BetaModels.ℙless — Method

ℙless(X::Beta{Float64}, Y::Beta{Float64}; δ = 0.0, method = "approx", n::Int = 10_000)

Return ℙ(X < Y + δ). The argument method may be one of:

"approx" (Normal approximation)

"montecarlo" (Monte-Carlo approximation)

"numeric" (quadrature).

If method = "montecarlo" then n indicates the number of samples to use.

BetaModels.ℙmax — Method

ℙmax(D::Vector{Beta{Float64}}; n = 10_000)

Probability that each Beta ∈ D is maximum via Monte Carlo simulation.

BetaModels.ℙmin — Method

ℙmin(D::Vector{Beta{Float64}}; n = 10_000)

Probability that each Beta ∈ D is minimum via Monte Carlo simulation.

Index

Distributions.Normal

Base.:-

Base.:-

BetaModels.ℙless

BetaModels.ℙmax

BetaModels.ℙmin